index/suffixarray: support for serialization
R=r
CC=golang-dev
https://golang.org/cl/5040041

+++ src/pkg/index/suffixarray/qsufsort.go
@@ -26,7 +26,7 @@ package suffixarray
 
 import "sort"
 
-func qsufsort(data []byte) []int {
+func qsufsort(data []byte) []int32 {
 	// initial sorting by first byte of suffix
 	sa := sortedByFirstByte(data)
 	if len(sa) < 2 {
@@ -39,20 +39,20 @@ func qsufsort(data []byte) []int {
 	// the index starts 1-ordered
 	sufSortable := &suffixSortable{sa, inv, 1}
 
-	for sa[0] > -len(sa) { // until all suffixes are one big sorted group
+	for int(sa[0]) > -len(sa) { // until all suffixes are one big sorted group
 		// The suffixes are h-ordered, make them 2*h-ordered
 		pi := 0 // pi is first position of first group
 		sl := 0 // sl is negated length of sorted groups
 		for pi < len(sa) {
-			if s := sa[pi]; s < 0 { // if pi starts sorted group
+			if s := int(sa[pi]); s < 0 { // if pi starts sorted group
 				pi -= s // skip over sorted group
 				sl += s // add negated length to sl
 			} else { // if pi starts unsorted group
 				if sl != 0 {
-					sa[pi+sl] = sl // combine sorted groups before pi
+					sa[pi+sl] = int32(sl) // combine sorted groups before pi
 					sl = 0
 				}
-				pk := inv[s] + 1 // pk-1 is last position of unsorted group
+				pk := int(inv[s]) + 1 // pk-1 is last position of unsorted group
 				sufSortable.sa = sa[pi:pk]
 				sort.Sort(sufSortable)
 				sufSortable.updateGroups(pi)
@@ -60,19 +60,19 @@ func qsufsort(data []byte) []int {
 			}
 		}
 		if sl != 0 { // if the array ends with a sorted group
-			sa[pi+sl] = sl // combine sorted groups at end of sa
+			sa[pi+sl] = int32(sl) // combine sorted groups at end of sa
 		}
 
 		sufSortable.h *= 2 // double sorted depth
 	}
 
 	for i := range sa { // reconstruct suffix array from inverse
-		sa[inv[i]] = i
+		sa[inv[i]] = int32(i)
 	}
 	return sa
 }
 
-func sortedByFirstByte(data []byte) []int {
+func sortedByFirstByte(data []byte) []int32 {
 	// total byte counts
 	var count [256]int
 	for _, b := range data {
@@ -84,17 +84,17 @@ func sortedByFirstByte(data []byte) []int {
 		count[b], sum = sum, count[b]+sum
 	}
 	// iterate through bytes, placing index into the correct spot in sa
-	sa := make([]int, len(data))
+	sa := make([]int32, len(data))
 	for i, b := range data {
-		sa[count[b]] = i
+		sa[count[b]] = int32(i)
 		count[b]++
 	}
 	return sa
 }
 
-func initGroups(sa []int, data []byte) []int {
+func initGroups(sa []int32, data []byte) []int32 {
 	// label contiguous same-letter groups with the same group number
-	inv := make([]int, len(data))
+	inv := make([]int32, len(data))
 	prevGroup := len(sa) - 1
 	groupByte := data[sa[prevGroup]]
 	for i := len(sa) - 1; i >= 0; i-- {
@@ -105,7 +105,7 @@ func initGroups(sa []int, data []byte) []int {
 			groupByte = b
 			prevGroup = i
 		}
-		inv[sa[i]] = prevGroup
+		inv[sa[i]] = int32(prevGroup)
 		if prevGroup == 0 {
 			sa[0] = -1
 		}
@@ -120,9 +120,9 @@ func initGroups(sa []int, data []byte) []int {
 			if data[sa[i]] == lastByte && s == -1 {
 				s = i
 			}
-			if sa[i] == len(sa)-1 {
+			if int(sa[i]) == len(sa)-1 {
 				sa[i], sa[s] = sa[s], sa[i]
-				inv[sa[s]] = s
+				inv[sa[s]] = int32(s)
 				sa[s] = -1 // mark it as an isolated sorted group
 				break
 			}
@@ -132,9 +132,9 @@ func initGroups(sa []int, data []byte) []int {
 }
 
 type suffixSortable struct {
-	sa  []int
-	inv []int
-	h   int
+	sa  []int32
+	inv []int32
+	h   int32
 }
 
 func (x *suffixSortable) Len() int           { return len(x.sa) }
@@ -156,7 +156,7 @@ func (x *suffixSortable) updateGroups(offset int) {
 	prev := 0
 	for _, b := range bounds {
 		for i := prev; i < b; i++ {
-			x.inv[x.sa[i]] = offset + b - 1
+			x.inv[x.sa[i]] = int32(offset + b - 1)
 		}
 		if b-prev == 1 {
 			x.sa[prev] = -1
